+++ packages/query/src/useMutationUserOperationCreate.ts
@@ -13,24 +13,28 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { updateWallet } from "@lightdotso/client";
-import type { WalletData } from "@lightdotso/data";
-import type { WalletParams } from "@lightdotso/params";
+import { createUserOperation } from "@lightdotso/client";
+import type { UserOperationCreateParams } from "@lightdotso/params";
 import { queryKeys } from "@lightdotso/query-keys";
+import type { UserOperation } from "@lightdotso/schemas";
 import { useAuth } from "@lightdotso/stores";
 import { toast } from "@lightdotso/ui";
 import { useQueryClient, useMutation } from "@tanstack/react-query";
+import type { Address } from "viem";
+import { toBytes, toHex } from "viem";
 
 // -----------------------------------------------------------------------------
 // Query Mutation
 // -----------------------------------------------------------------------------
 
-export const useMutationUserOperationCreate = (params: WalletParams) => {
+export const useMutationUserOperationCreate = (
+  params: UserOperationCreateParams,
+) => {
   // ---------------------------------------------------------------------------
   // Stores
   // ---------------------------------------------------------------------------
 
-  const { clientType } = useAuth();
+  const { clientType, address } = useAuth();
 
   // ---------------------------------------------------------------------------
   // Query
@@ -39,18 +43,57 @@ export const useMutationUserOperationCreate = (params: WalletParams) => {
   const queryClient = useQueryClient();
 
   const { mutate, isPending, isSuccess, isError } = useMutation({
-    mutationFn: async (data: Partial<WalletData>) => {
-      const loadingToast = toast.loading("Updating name...");
+    mutationFn: async (data: Partial<UserOperation>) => {
+      if (
+        !data.chainId ||
+        !data.hash ||
+        data.nonce === undefined ||
+        data.nonce === null ||
+        !data.initCode ||
+        !data.sender ||
+        !data.callData ||
+        !data.callGasLimit ||
+        !data.verificationGasLimit ||
+        !data.preVerificationGas ||
+        !data.maxFeePerGas ||
+        !data.maxPriorityFeePerGas ||
+        !data.paymasterAndData
+      ) {
+        return;
+      }
 
-      const res = await updateWallet(
+      const loadingToast = toast.loading("Creating the transaction...");
+
+      // Replace with your actual fetch logic
+      const res = await createUserOperation(
         {
           params: {
             query: {
-              address: params.address,
+              chain_id: Number(data.chainId),
             },
           },
           body: {
-            name: data.name,
+            signature: {
+              owner_id: params.ownerId,
+              signature: toHex(
+                new Uint8Array([...toBytes(params.signedData), 2]),
+              ),
+              signature_type: 1,
+            },
+            user_operation: {
+              chain_id: Number(data.chainId),
+              hash: data.hash,
+              nonce: Number(data.nonce),
+              init_code: data.initCode,
+              sender: data.sender,
+              call_data: data.callData,
+              call_gas_limit: Number(data.callGasLimit),
+              verification_gas_limit: Number(data.verificationGasLimit),
+              pre_verification_gas: Number(data.preVerificationGas),
+              max_fee_per_gas: Number(data.maxFeePerGas),
+              max_priority_fee_per_gas: Number(data.maxPriorityFeePerGas),
+              paymaster_and_data: data.paymasterAndData,
+            },
           },
         },
         clientType,
@@ -61,13 +104,13 @@ export const useMutationUserOperationCreate = (params: WalletParams) => {
       // Return if the response is 200
       res.match(
         _ => {
-          toast.success("Successfully updated name.");
+          toast.success("Successfully created transaction!");
         },
         err => {
           if (err instanceof Error) {
             toast.error(err.message);
           } else {
-            toast.error("Failed to update name.");
+            toast.error("Failed to create transaction.");
           }
 
           throw err;
@@ -75,49 +118,28 @@ export const useMutationUserOperationCreate = (params: WalletParams) => {
       );
     },
     // When mutate is called:
-    onMutate: async (wallet: Partial<WalletData>) => {
-      // Cancel any outgoing refetches
-      // (so they don't overwrite our optimistic update)
-      await queryClient.cancelQueries({
-        queryKey: queryKeys.wallet.get({ address: params.address }).queryKey,
-      });
+    // onMutate: async (wallet: WalletCreateParams) => {
+    // Cancel any outgoing refetches
+    // (so they don't overwrite our optimistic update)
+    // await queryClient.cancelQueries({
+    // queryKey: queryKeys.user_operation.list({ address: address as Address }).queryKey,
+    // });
 
-      // Snapshot the previous value
-      const previousSettings: WalletData | undefined = queryClient.getQueryData(
-        queryKeys.wallet.get({ address: params.address }).queryKey,
-      );
-
-      // Optimistically update to the new value
-      queryClient.setQueryData(
-        queryKeys.wallet.settings({ address: params.address }).queryKey,
-        (old: WalletData) => {
-          return { ...old, wallet };
-        },
-      );
-
-      // Return a context object with the snapshotted value
-      return { previousSettings };
-    },
+    // return { wallet };
+    // },
     // If the mutation fails, use the context we returned above
-    onError: (err, _newWalletSettings, context) => {
-      queryClient.setQueryData(
-        queryKeys.wallet.get({ address: params.address }).queryKey,
-        context?.previousSettings,
-      );
-
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    onError: (err, _newWalletSettings, _context) => {
       if (err instanceof Error) {
         toast.error(err.message);
       } else {
-        toast.error("Failed to update name.");
+        toast.error("Failed to create transaction.");
       }
     },
     onSettled: () => {
       queryClient.invalidateQueries({
-        queryKey: queryKeys.wallet.get({ address: params.address }).queryKey,
-      });
-      queryClient.invalidateQueries({
         queryKey: queryKeys.wallet.list({
-          address: params.address,
+          address: address as Address,
           limit: Number.MAX_SAFE_INTEGER,
           offset: 0,
         }).queryKey,
@@ -126,7 +148,7 @@ export const useMutationUserOperationCreate = (params: WalletParams) => {
       // Invalidate the cache for the address
       // fetch(`/api/revalidate/tag?tag=${address}`);
     },
-    mutationKey: queryKeys.wallet.get({ address: params.address }).queryKey,
+    // mutationKey: queryKeys.wallet.create({ address }).queryKey,
   });
 
   return {
